requirements.txt
bloscpack==0.16.0
CommonModules==0.1.26
CommonModules==0.1.26
datatable==1.1.0
dgl_cu117==0.9.1.post1
icecream==2.1.3
ipdb==0.13.13
numpy==1.24.4
pandas==1.5.2
pytorch_lightning==2.1.0
scikit_learn==1.1.3
scipy==1.9.1
sentence_transformers==2.2.2
torch==2.1.0
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.34.0
